test: provider-aware database test architecture
The suite now selects its engine, and running it no longer depends on any one
database being reachable.

Test configuration is resolved in tests/support/database.py and never falls
back to the application's variables. That separation is the point: a suite that
can read KAE_DATABASE_URL is one mistaken environment away from truncating a
real database, which has already happened here once, when Alembic resolved that
variable ahead of the throwaway database a fixture had built.

The destructive guard now checks the database name rather than the host. The
old version accepted anything on localhost, which is every database on a
developer's machine including the one holding the corpus. A target must name
itself disposable — _test, test_, testing — and known application names are
refused outright. Its tests moved to tests/unit, where they run without a
database, because the check protecting the other tests must not depend on the
thing it protects them from.

PostgreSQL is the default test provider. A default engine is safe; a default
URL would not be, so there is none: without one, live-database tests skip with
a reason and `pytest` still runs. 208 tests pass with no database configured at
all.

Tests are marked by the fixtures they request rather than by hand. Six hundred
existing tests would otherwise need annotating, and a marker people forget is
worse than one derived from what a test actually asks for.

CockroachDB is selectable, not deleted. Its tests skip with a reason naming the
selected provider, and CI runs it as a separate job gated on a variable so the
pipeline passes without a cluster. Nothing was commented out or removed.

Also splits kae_memory/config.py out of persistence/providers.py. Provider
selection is configuration; which DDL a provider compiles is persistence. The
first version put both below the repository boundary, so the MCP adapter
imported persistence to read a URL and broke the one-way dependency rule that
ADR-0018 enforces with a test. The test was right.

+++ src/kae_memory/persistence/tables.py
@@ -111,7 +111,11 @@ class KnowledgeVersionRow(Base):
     __tablename__ = "knowledge_versions"
     __table_args__ = (UniqueConstraint("knowledge_item_id", "version_number"),)
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    #: 64-bit deliberately. ``Integer`` compiles to a 32-bit column on
+    #: PostgreSQL and a 64-bit one on CockroachDB, whose INT is INT8 — so the
+    #: same migration produced two different schemas, and CockroachDB's
+    #: ``unique_rowid()`` values do not fit in the narrower one.
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     knowledge_item_id: Mapped[str] = mapped_column(
         ForeignKey("knowledge_items.id", ondelete="CASCADE"), nullable=False, index=True
     )
